Finish sphere intersection from 3.2.3 to 3.2.6
Also, split the common logic of intersect_p in order to not
duplicate it and keep it as a private member function of Sphere.
Change the tests that are already written to test intersect_p instead.

That means we aren't completely done -- we still need to test that the
differential geometry information is populated correctly for intersection
points...