The great shape refactor
Previously, we were using shape as a trait when in reality, it's an enum.
The ADT formulation allows all shapes to share a common interface, and we
can default to their given modules whenever we need to do our processing.

Dynamic dispatch is better suited to this approach, in my opinion...