The great primitive refactor.
We finally connected shapes to the rest of the infrastructure. This
meant two major additions to the design of the ray tracer:

1. Scenes, renderers, integrators, and primitives all share a
lifetime parameter. This may need to be revisited, I think there
may be instances when these things don't outlive each other, but
I'm not totally sure what those instances are.

2. We need to use atomically counted references in the loop subdivision
surface implementation. This shouldn't cause any problems except
performance issues (although it's really Rust being pedantic, if we had
a better way of storing the mesh than a vector of Rc's then we wouldn't
run into this issue.)