Very large commit (I feel bad about this)
- Finished first draft of grid accelerator. Still needs a test to make sure we can intersect
primitives within it.

- Changed primitives to store their ID as a base and retain their implementation behind
an atomically shared reference. This way we can share our primitive objects with Intersections
without having to pass around lifetimes and annoy the compiler

- Removed lifetime restrictions on anything having to do with primitives or shapes like
differential geometry and bsdfs. This will increase the size of the intersection structure, but
if it becomes a problem we can encapsulate the shape base in the same way we do with the primitive
base.