Add partition_by function.
This function partially sorts a mutable slice such that the values
in the bottom half of the array are less than the values in the top
half of the array. The definition of "less than" is determined by
the function passed to partition_by.

This function is useful in building BVH trees, and should be replaced
with rust-lang/rfcs#1470 if it matures into the stdlib.